Compatibility update
+ Moved build system from the deprecated numpy distutils and easy_install to cython and python build.
+ Moved from python 3.9 to 3.10.
+ Moved from blender version <2.8 to 3.6, it is no longer necessary to compile manually.
+ Tidied up the examples to be more readable for new users.

+++ examples/02_NMNIST/1_saccades_NMNIST.py
@@ -2,7 +2,7 @@
     Simulate saccadic movement along several axis directions in front of One NMNIST digit
 """
 import bpy
-import mathutils
+from mathutils import Vector
 import cv2
 import math
 import numpy as np
@@ -17,14 +17,14 @@ from tqdm import tqdm
 
 path_NMNIST = os.path.abspath(os.getcwd()) + "/../../data/"
 
-if not os.path.exists(path_NMNIST + "text"):
-    os.mkdir(path_NMNIST + "text")
-if not os.path.exists(path_NMNIST + "labels"):
-    os.mkdir(path_NMNIST + "labels")
 if not os.path.exists(path_NMNIST + "tmp"):
     os.mkdir(path_NMNIST + "tmp")
-if not os.path.exists(path_NMNIST + "events"):
-    os.mkdir(path_NMNIST + "events")
+if not os.path.exists(path_NMNIST + "tmp/text"):
+    os.mkdir(path_NMNIST + "tmp/text")
+if not os.path.exists(path_NMNIST + "tmp/labels"):
+    os.mkdir(path_NMNIST + "tmp/labels")
+if not os.path.exists(path_NMNIST + "tmp/events"):
+    os.mkdir(path_NMNIST + "tmp/events")
 
 text_start = 0
 text_end = 1
@@ -33,48 +33,61 @@ if not os.path.exists(path_NMNIST + 'tmp/tmp_{}_{}'.format(text_start, text_end)
     os.mkdir(path_NMNIST + 'tmp/tmp_{}_{}'.format(text_start, text_end))
 
 scene = bpy.context.scene
+
 bpy.ops.object.select_all(action='DESELECT')
-bpy.data.objects['Camera'].select = True
-bpy.ops.object.delete()
-bpy.data.objects['Cube'].select = True
-bpy.ops.object.delete()
-bpy.data.objects['Lamp'].location = mathutils.Vector((5, 5, 5))
-bpy.data.objects['Lamp'].data.energy = 10
-bpy.data.objects['Lamp'].data.type = 'HEMI'
 
-# Create background
-offset = 100  # Background light
-bpy.ops.mesh.primitive_plane_add(radius=10.0, calc_uvs=True, view_align=False, enter_editmode=False,
-                                 location=(-10.0 + text_start * 20, 0.0, 5.0), rotation=(0, 0, 0))
+if "Camera" in bpy.data.objects:
+    bpy.data.objects['Camera'].select_set(True)
+    bpy.ops.object.delete()
+
+if "Cube" in bpy.data.objects:
+    bpy.data.objects['Cube'].select_set(True)
+    bpy.ops.object.delete()
+
+if "Light" in bpy.data.objects:
+    light = bpy.data.objects['Light']
+    light.location = Vector((5, 5, 5))
+    light.data.energy = 10
+
+    light.data.type = 'SUN'
+
+offset = 100
+
+text_start = 0
+bpy.ops.mesh.primitive_plane_add(size=20.0, calc_uvs=True, enter_editmode=False,
+                                 location=(10 - 20.0 * text_start, 0.0, 5.0), rotation=(0, 0, 0))
 pl = bpy.context.active_object
-pl.name = 'PlaneSky'
-mat = bpy.data.materials.new('Sky')
-tx = bpy.data.textures.new('SkyMap', 'IMAGE')
-tx.extension = 'REPEAT'
-img_texture = np.full([2800, 2800], 100)
-cv2.imwrite(path_NMNIST + "img_text.png", img_texture)
-im = bpy.data.images.load(path_NMNIST + "img_text.png")
-bpy.data.textures['SkyMap'].image = im
-bpy.data.materials['Sky'].texture_slots.add()
-bpy.data.materials['Sky'].texture_slots[0].texture = tx
-pl.data.materials.append(bpy.data.materials['Sky'])
+pl.name = 'NMNIST_PLANE'
 
-# Generate the annimation for a given sequence
-t = 0
-bpy.ops.mesh.primitive_plane_add(radius=10.0, calc_uvs=True, view_align=False, enter_editmode=False,
-                                 location=(10 + 20.0 * t, 0.0, 5.0), rotation=(0, 0, 0))
-pl = bpy.context.active_object
-pl.name = 'PlaneSky_{}'.format(t)
-mat = bpy.data.materials.new('Sky_{}'.format(t))
-tx = bpy.data.textures.new('SkyMap_{}'.format(t), 'IMAGE')
-tx.extension = 'REPEAT'
-im = bpy.data.images.load(path_NMNIST + 'text/img_temp_{}.png'.format(t))
-bpy.data.textures['SkyMap_{}'.format(t)].image = im
-bpy.data.materials['Sky_{}'.format(t)].texture_slots.add()
-bpy.data.materials['Sky_{}'.format(t)].texture_slots[0].texture = tx
-pl.data.materials.append(bpy.data.materials['Sky_{}'.format(t)])
+material = bpy.data.materials.new(name='NMNIST_PLANE_MATERIAL')
+material.use_nodes = True
+nodes = material.node_tree.nodes
 
-# Create the camera
+for node in nodes:
+    nodes.remove(node)
+
+texture = bpy.data.textures.new('NMNIST_PLANE_TEX', 'IMAGE')
+texture.extension = 'REPEAT'
+
+texture.image = bpy.data.images.load(path_NMNIST + 'tmp/text/mnist_plane_texture.png')
+
+shader_node = nodes.new(type='ShaderNodeBsdfPrincipled')
+shader_node.location = (0, 0)
+
+texture_node = nodes.new(type='ShaderNodeTexImage')
+texture_node.location = (-300, 0)
+texture_node.image = texture.image
+
+material_output_node = nodes.new(type='ShaderNodeOutputMaterial')
+material_output_node.location = (300, 0)
+
+links = material.node_tree.links
+link = links.new(texture_node.outputs["Color"], shader_node.inputs["Base Color"])
+
+link = links.new(shader_node.outputs["BSDF"], material_output_node.inputs["Surface"])
+
+pl.data.materials.append(material)
+
 ppsee = Blender_DvsSensor("PropheseeSensor")
 ppsee.set_shape(28, 28)
 ppsee.set_sensor(28, 28, 0.015)
@@ -82,17 +95,19 @@ ppsee.set_dvs_sensor(th_pos=0.15, th_neg=0.15, th_n=0.05, lat=100, tau=300, jit=
 ppsee.init_tension()
 ppsee.init_bgn_hist("../../data/noise_pos_0.1lux.npy", "../../data/noise_neg_0.1lux.npy")
 ppsee.set_sensor_optics(20)
-scene.objects.link(ppsee.cam)
+scene = bpy.context.scene
+master_collection = bpy.context.collection
+
+master_collection.objects.link(ppsee.cam)
 scene.camera = ppsee.cam
 ppsee.set_angle([math.pi, 0, 0.0])
 ppsee.set_position([0.0, 0.0, 0.0])
 ppsee.init_tension()
 
 scene.render.image_settings.file_format = 'PNG'
-scene.render.filepath = path_NMNIST + 'tmp/tmp_{}_{}/image_annim'.format(text_start, text_end)
-scene.render.resolution_x = 2 * ppsee.def_x
-scene.render.resolution_y = 2 * ppsee.def_y
-#scene.render.engine = 'CYCLES'
+scene.render.filepath = path_NMNIST + 'tmp/tmp_{}_{}/image_anim'.format(text_start, text_end)
+scene.render.resolution_x = ppsee.def_x
+scene.render.resolution_y = ppsee.def_y
 
 ed = EventDisplay("Events", 28, 28, 2000)
 ed.display_time = False
@@ -107,7 +122,7 @@ speed = 0.01
 t = 0
 pos_img = 0
 dt = 1000  # us
-labels = np.load(path_NMNIST + "text/NMNIST_lab.npy")
+labels = np.load(path_NMNIST + "tmp/text/NMNIST_lab.npy")
 lab = np.zeros((2, len(labels)))  # [ ts, lab]
 bpy.context.scene.frame_set(pos_img)
 ppsee.set_position([pos[0], pos[1], pos[2]])
@@ -151,6 +166,11 @@ for nb_digit in tqdm(range(0, 10, 1)):
 
 bpy.context.scene.frame_end = pos_img
 bpy.context.scene.frame_current = bpy.context.scene.frame_start
+
+# Specify the filepath for the .blend file
+filepath = path_NMNIST + "/tmp/nmnist_experiment.blend"
+bpy.ops.wm.save_as_mainfile(filepath=filepath)
+
 bpy.ops.render.render(animation=True)
 
 # Events
@@ -159,7 +179,7 @@ fourcc = cv2.VideoWriter_fourcc('M', 'J', 'P', 'G')
 out = cv2.VideoWriter('nmnist.avi', fourcc, 20.0, (ppsee.def_x, ppsee.def_y))
 for i in range(1, pos_img, 1):
     t = t + dt
-    filename = path_NMNIST + 'tmp/tmp_{}_{}/'.format(text_start, text_end) + 'image_annim{0:04}.png'.format(i)
+    filename = path_NMNIST + 'tmp/tmp_{}_{}/'.format(text_start, text_end) + 'image_anim{0:04}.png'.format(i)
     im = cv2.imread(filename)
     out.write(im)
     if t == dt:
@@ -171,5 +191,5 @@ for i in range(1, pos_img, 1):
     cv2.imshow("Blender", im)
     cv2.waitKey(10)
 out.release()
-ev.write(path_NMNIST + 'events/MNIST_ev_{}_{}_td.dat'.format(text_start, text_end))
-np.save(path_NMNIST + 'labels/MNIST_lab_events_{}_{}.npy'.format(text_start, text_end), lab)
+ev.write(path_NMNIST + 'tmp/events/MNIST_ev_{}_{}_td.dat'.format(text_start, text_end))
+np.save(path_NMNIST + 'tmp/labels/MNIST_lab_events_{}_{}.npy'.format(text_start, text_end), lab)
